refactor: rename file to match server
