+++ client/src/services/index.js
@@ -1,6 +1,6 @@
 export * from './participant';
 export * from './psi-cohort';
 export * from './return-of-service';
-export * from './feature-flag';
+export * from './feature-flags';
 export * from './notifications';
 export * from './participant-details';
